% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour.R
\name{adjust}
\alias{adjust}
\title{Adjust colour palette}
\usage{
adjust(colours, palette, n = length(colours))
}
\arguments{
\item{colours}{A \code{\link{character}} vector of colours.}

\item{palette}{A \code{\link{character}} string giving the name of
the palette to be used (see \code{\link{colour}}).}

\item{n}{A \code{\link{numeric}} scalar giving the number of colours to
select.}
}
\value{
A character vector of colours.
}
\description{
Adjusts a color palette according to specific rules.
}
\details{
If there is no rule for the palette, \code{n} colours are
 randomly selected.
}
\author{
N. Frerebeau
}
