% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkPackage}
\alias{checkPackage}
\title{Check package import}
\usage{
checkPackage(package)
}
\arguments{
\item{package}{A \code{\link{character}} string giving the name of
the needed package.}
}
\description{
Checks if the required package is available.
}
\details{
An error is signaled if the required package is not available.
}
\author{
N. Frerebeau
}
