% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{Colour Difference}
\usage{
compare(x, metric = 2000, diag = FALSE, upper = FALSE)
}
\arguments{
\item{x}{A \code{\link{character}} vector of colours.}

\item{metric}{An \code{\link{integer}} value giving the year the metric was
recommended by the CIE. It must be one of "\code{1976}", "\code{1994}", or
"\code{2000}".}

\item{diag}{A \code{\link{logical}} scalar: should the diagonal of the
distance matrix be printed?}

\item{upper}{A \code{\link{logical}} scalar: should the upper triangle of the
distance matrix should be printed?}
}
\value{
A \code{\link[stats:dist]{distance matrix}}.
}
\description{
Computes CIELAB distance metric (\eqn{\DeltaE*}).
}
\examples{
# Trichromat
pal <- colour("bright")

(deltaE <- compare(pal(5)))
summary(deltaE)

# Deuteranopia
deu <- convert(pal, mode = "deuteranopia")
compare(deu(5))

# Protanopia
pro <- convert(pal, mode = "protanopia")
compare(pro(5))

# Tritanopia
tri <- convert(pal, mode = "tritanopia")
compare(tri(5))

# Achromatopsia
ach <- convert(pal, mode = "achromatopsia")
compare(ach(5))
}
\seealso{
Other diagnostic tools: 
\code{\link{convert}()},
\code{\link{plot}()}
}
\author{
N. Frerebeau
}
\concept{diagnostic tools}
