% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{crameri}
\alias{crameri}
\title{Fabio Crameri's Color Schemes}
\source{
Crameri, F. (2023). Scientific colour maps. \emph{Zenodo}, v8.0.1.
\doi{10.5281/zenodo.1243862}
}
\description{
Fabio Crameri's Color Schemes
}
\section{Diverging Color Schemes}{

If more colors than defined are needed from a given scheme, the color
coordinates are linearly interpolated to provide a continuous version of the
scheme.

\tabular{ll}{
\strong{Palette} \tab \strong{Max.} \cr
\code{bam}       \tab 256      \cr
\code{bamO}*     \tab 256      \cr
\code{berlin}    \tab 256      \cr
\code{broc}      \tab 256      \cr
\code{brocO}*    \tab 256      \cr
\code{cork}      \tab 256      \cr
\code{corkO}*    \tab 256      \cr
\code{lisbon}    \tab 256      \cr
\code{managua}   \tab 256      \cr
\code{roma}      \tab 256      \cr
\code{romaO}*    \tab 256      \cr
\code{tofino}    \tab 256      \cr
\code{vanimo}    \tab 256      \cr
\code{vik}       \tab 256      \cr
\code{vikO}*     \tab 256      \cr
}
*: cyclic color schemes.
}

\section{Sequential Color Schemes}{

If more colors than defined are needed from a given scheme, the color
coordinates are linearly interpolated to provide a continuous version of the
scheme.

\tabular{ll}{
\strong{Palette} \tab \strong{Max.} \cr
\code{acton}     \tab 256      \cr
\code{bamako}    \tab 256      \cr
\code{batlow}    \tab 256      \cr
\code{batlowW}   \tab 256      \cr
\code{batlowK}   \tab 256      \cr
\code{bilbao}    \tab 256      \cr
\code{buda}      \tab 256      \cr
\code{bukavu}*   \tab 256      \cr
\code{davos}     \tab 256      \cr
\code{devon}     \tab 256      \cr
\code{fes}*      \tab 256      \cr
\code{glasgow}   \tab 256      \cr
\code{grayC}     \tab 256      \cr
\code{hawaii}    \tab 256      \cr
\code{imola}     \tab 256      \cr
\code{lajolla}   \tab 256      \cr
\code{lapaz}     \tab 256      \cr
\code{navia}     \tab 256      \cr
\code{naviaW}    \tab 256      \cr
\code{nuuk}      \tab 256      \cr
\code{oleron}*   \tab 256      \cr
\code{oslo}      \tab 256      \cr
\code{tokyo}     \tab 256      \cr
\code{turku}     \tab 256      \cr
}
*: multisequential color schemes.
}

\references{
Crameri, F. (2018). Geodynamic diagnostics, scientific visualisation and
StagLab 3.0. \emph{Geosci. Model Dev.}, 11, 2541-2562.
\doi{10.5194/gmd-11-2541-2018}

Crameri, F., Shephard, G. E. & Heron, P. J. (2020). The misuse of colour in
science communication. \emph{Nature Communications}, 11, 5444.
\doi{10.1038/s41467-020-19160-7}
}
\keyword{internal}
