% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/khroma-package.R
\docType{package}
\name{khroma-package}
\alias{khroma-package}
\alias{khroma}
\title{khroma: Colour Schemes for Scientific Data Visualization}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Color schemes ready for each type of data (qualitative, diverging or sequential), with colors that are distinct for all people, including color-blind readers. This package provides an implementation of Paul Tol (2018) and Fabio Crameri (2018) \doi{10.5194/gmd-11-2541-2018} color schemes for use with 'graphics' or 'ggplot2'. It provides tools to simulate color-blindness and to test how well the colors of any palette are identifiable. Several scientific thematic schemes (geologic timescale, land cover, FAO soils, etc.) are also implemented.
}
\details{
\tabular{ll}{
\strong{Version} \tab 1.15.0 \cr
\strong{License} \tab GPL-3 \cr
\strong{CRAN DOI} \tab \doi{10.32614/CRAN.package.khroma} \cr
\strong{Zenodo DOI} \tab \doi{10.5281/zenodo.1472077} \cr
}

Archéosciences Bordeaux (UMR 6034)\cr
Maison de l'Archéologie\cr
Université Bordeaux Montaigne\cr
F-33607 Pessac cedex\cr
France
}
\seealso{
Useful links:
\itemize{
  \item \url{https://codeberg.org/tesselle/khroma}
  \item \url{https://tesselle.r-universe.dev/khroma}
  \item \url{https://packages.tesselle.org/khroma/}
  \item Report bugs at \url{https://codeberg.org/tesselle/khroma/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Frerebeau \email{nicolas.frerebeau@u-bordeaux-montaigne.fr} (\href{https://orcid.org/0000-0001-5759-4944}{ORCID})

Other contributors:
\itemize{
  \item Brice Lebrun (\href{https://orcid.org/0000-0001-7503-8685}{ORCID}) (Logo designer) [artist]
  \item Vincent Arel-Bundock \email{vincent.arel-bundock@umontreal.ca} (\href{https://orcid.org/0000-0003-2042-7063}{ORCID}) [contributor]
  \item Ulrik Stervbo \email{ulrik.stervbo@gmail.com} (\href{https://orcid.org/0000-0002-2831-8868}{ORCID}) [contributor]
  \item Université Bordeaux Montaigne (03pbgwk21) [funder]
  \item CNRS (02feahw73) [funder]
}

}
\keyword{internal}
