% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.R
\name{scale}
\alias{scale}
\title{Colour scale constructor}
\usage{
scale(aesthetics, palette, reverse = FALSE, ...)
}
\arguments{
\item{aesthetics}{The names of the aesthetics that this scale works with.}

\item{palette}{A \code{\link{character}} string giving the name of
the palette to be used (see \code{\link{colour}}).}

\item{reverse}{A \code{\link{logical}} scalar specifying if the resulting
vector of colours should be reversed.}

\item{...}{Further arguments passed to \code{\link[ggplot2]{discrete_scale}}
or \code{\link[ggplot2]{continuous_scale}}, used respectively for
qualitative data and diverging/sequential data.}
}
\description{
Builds discrete or continuous scale for ggplot2 according to the color
 scheme used.
}
\author{
N. Frerebeau
}
