% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colour_crameri.R
\name{scale_crameri_bamO}
\alias{scale_crameri_bamO}
\alias{scale_colour_bamO}
\alias{scale_color_bamO}
\alias{scale_fill_bamO}
\title{Fabio Crameri's \emph{bamO} Cyclic Color Scheme}
\source{
Crameri, F. (2023). Scientific colour maps. \emph{Zenodo}, v8.0.1.
\doi{10.5281/zenodo.1243862}
}
\usage{
scale_colour_bamO(
  ...,
  reverse = FALSE,
  range = c(0, 1),
  discrete = FALSE,
  aesthetics = "colour"
)

scale_color_bamO(
  ...,
  reverse = FALSE,
  range = c(0, 1),
  discrete = FALSE,
  aesthetics = "colour"
)

scale_fill_bamO(
  ...,
  reverse = FALSE,
  range = c(0, 1),
  discrete = FALSE,
  aesthetics = "fill"
)
}
\arguments{
\item{...}{Arguments passed to \code{\link[ggplot2:continuous_scale]{ggplot2::continuous_scale()}}.}

\item{reverse}{A \code{\link{logical}} scalar. Should the resulting
vector of colors be reversed?}

\item{range}{A length-two \code{\link{numeric}} vector specifying the
fraction of the scheme's color domain to keep.}

\item{discrete}{A \code{\link{logical}} scalar: should the color scheme be
used as a discrete scale?}

\item{aesthetics}{A \code{\link{character}} string or vector of character
strings listing the name(s) of the aesthetic(s) that this scale works with.}
}
\value{
A \link[ggplot2:continuous_scale]{continuous} scale.
}
\description{
Fabio Crameri's \emph{bamO} Cyclic Color Scheme
}
\section{Diverging Color Schemes}{

If more colors than defined are needed from a given scheme, the color
coordinates are linearly interpolated to provide a continuous version of the
scheme.

\tabular{ll}{
\strong{Palette} \tab \strong{Max.} \cr
\code{bam}       \tab 256      \cr
\code{bamO}*     \tab 256      \cr
\code{berlin}    \tab 256      \cr
\code{broc}      \tab 256      \cr
\code{brocO}*    \tab 256      \cr
\code{cork}      \tab 256      \cr
\code{corkO}*    \tab 256      \cr
\code{lisbon}    \tab 256      \cr
\code{managua}   \tab 256      \cr
\code{roma}      \tab 256      \cr
\code{romaO}*    \tab 256      \cr
\code{tofino}    \tab 256      \cr
\code{vanimo}    \tab 256      \cr
\code{vik}       \tab 256      \cr
\code{vikO}*     \tab 256      \cr
}
*: cyclic color schemes.
}

\examples{
data(economics, package = "ggplot2")

ggplot2::ggplot(economics, ggplot2::aes(psavert, pce, colour = unemploy)) +
  ggplot2::geom_point() +
  scale_colour_broc(reverse = TRUE, midpoint = 12000)

ggplot2::ggplot(economics, ggplot2::aes(psavert, pce, colour = unemploy)) +
  ggplot2::geom_point() +
  scale_colour_berlin(midpoint = 9000)
}
\references{
Crameri, F. (2018). Geodynamic diagnostics, scientific visualisation and
StagLab 3.0. \emph{Geosci. Model Dev.}, 11, 2541-2562.
\doi{10.5194/gmd-11-2541-2018}

Crameri, F., Shephard, G. E. & Heron, P. J. (2020). The misuse of colour in
science communication. \emph{Nature Communications}, 11, 5444.
\doi{10.1038/s41467-020-19160-7}
}
\seealso{
Other cyclic color schemes: 
\code{\link{scale_crameri_brocO}},
\code{\link{scale_crameri_corkO}},
\code{\link{scale_crameri_romaO}},
\code{\link{scale_crameri_vikO}}

Other Fabio Crameri's color schemes: 
\code{\link{scale_crameri_acton}},
\code{\link{scale_crameri_bam}},
\code{\link{scale_crameri_bamako}},
\code{\link{scale_crameri_batlow}},
\code{\link{scale_crameri_batlowK}},
\code{\link{scale_crameri_batlowW}},
\code{\link{scale_crameri_berlin}},
\code{\link{scale_crameri_bilbao}},
\code{\link{scale_crameri_broc}},
\code{\link{scale_crameri_brocO}},
\code{\link{scale_crameri_buda}},
\code{\link{scale_crameri_bukavu}},
\code{\link{scale_crameri_cork}},
\code{\link{scale_crameri_corkO}},
\code{\link{scale_crameri_davos}},
\code{\link{scale_crameri_devon}},
\code{\link{scale_crameri_fes}},
\code{\link{scale_crameri_glasgow}},
\code{\link{scale_crameri_grayC}},
\code{\link{scale_crameri_hawaii}},
\code{\link{scale_crameri_imola}},
\code{\link{scale_crameri_lajolla}},
\code{\link{scale_crameri_lapaz}},
\code{\link{scale_crameri_lipari}},
\code{\link{scale_crameri_lisbon}},
\code{\link{scale_crameri_managua}},
\code{\link{scale_crameri_navia}},
\code{\link{scale_crameri_naviaW}},
\code{\link{scale_crameri_nuuk}},
\code{\link{scale_crameri_oleron}},
\code{\link{scale_crameri_oslo}},
\code{\link{scale_crameri_roma}},
\code{\link{scale_crameri_romaO}},
\code{\link{scale_crameri_tofino}},
\code{\link{scale_crameri_tokyo}},
\code{\link{scale_crameri_turku}},
\code{\link{scale_crameri_vanimo}},
\code{\link{scale_crameri_vik}},
\code{\link{scale_crameri_vikO}}
}
\author{
N. Frerebeau
}
\concept{Fabio Crameri's color schemes}
\concept{cyclic color schemes}
