% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{science}
\alias{science}
\title{Scientific Color Schemes}
\description{
Scientific Color Schemes
}
\details{
The following (qualitative) color schemes are available:
\describe{
\item{\code{stratigraphy}}{International Chronostratigraphic Chart (175 colors).}
\item{\code{land}}{AVHRR Global Land Cover Classification (14 colors).}
\item{\code{soil}}{FAO Reference Soil Groups (24 colors).}
}
}
\references{
Jones, A., Montanarella, L. & Jones, R. (Ed.) (2005). \emph{Soil atlas of
Europe}. Luxembourg: European Commission, Office for Official Publications
of the European Communities. 128 pp. ISBN: 92-894-8120-X.

Tol, P. (2021). \emph{Colour Schemes}. SRON. Technical Note No.
SRON/EPS/TN/09-002, issue 3.2.
URL: \url{https://sronpersonalpages.nl/~pault/data/colourschemes.pdf}

\href{https://www.ccgm.org/}{Commission for the Geological Map of the World}.
}
\keyword{internal}
