% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactoMineR.R, R/FactoMineR-PCA.R
\name{visualize}
\alias{visualize}
\alias{visualize.PCA}
\title{Data visualization}
\usage{
\method{visualize}{PCA}(x, axes = c(1, 2), map = c("individuals",
  "variables", "eigenvalues", "variance"), extra = NULL, group = NULL,
  ...)
}
\arguments{
\item{x}{An object of class \code{\link[FactoMineR]{PCA}}}

\item{axes}{A length-two \code{\link{numeric}} vector specifying the
components to plot.}

\item{map}{A \code{\link{character}} string or vector of strings specifying
the graph to plot. This must be one of "\code{individuals}",
"\code{variables}", "\code{eigenvalues}" or "\code{variance}".
Any unambiguous substring can be given.}

\item{extra}{A string or vector of \code{\link{character}} strings
specifiying the supplementary data to plot.
This must be one or more of "\code{individuals}", "\code{qualitative}",
"\code{quantitative}". Any unambiguous substring can be given.}

\item{group}{A \code{\link{character}} vector of categories or a length-one
\code{\link{numeric}} vector specifiying the categorial variable from which
to color the individuals.
The elements are coerced to characters by \code{\link{as.character}}.}

\item{...}{Further arguments passed to other methods.}
}
\description{
Visualize multivariate exploratory data analysis results.
}
\examples{
library(FactoMineR)
library(ggplot2)

# PCA
data(decathlon)
results <- PCA(decathlon, quanti.sup = 11:12, quali.sup = 13, graph = FALSE)
## Plot individuals and draw ellipses
visualize(results, map = "ind", axes = c(1, 2), group = 13) +
  ggplot2::stat_ellipse() +
  khroma::scale_color_bright()
## Plot variables with labels
visualize(results, map = "variables", axes = c(1, 2)) +
  ggplot2::geom_label()
## Plot supplementary variables
visualize(results, map = "variables", axes = c(1, 2),
          extra = c("quanti", "quali")) +
  ggplot2::geom_label()
## Plot eigenvalues
visualize(results, map = "eig", axes = c(1, 2))
## Plot cumulative percentage of variance
visualize(results, map = "variance", axes = c(1, 2))
}
\seealso{
\link[FactoMineR]{PCA}, \link[FactoMineR]{CA}
}
\author{
N. Frerebeau
}
