% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{info}
\alias{info}
\title{Available Schemes}
\usage{
info()
}
\value{
A \code{\link{data.frame}} with the following columns:
\describe{
\item{\code{palette}}{Names of palette.}
\item{\code{type}}{Types of schemes: sequential, diverging or qualitative.}
\item{\code{max}}{Maximum number of colors that are contained in each
palette. Only relevant for qualitative schemes.}
\item{\code{missing}}{The hexadecimal color value for mapping missing values.}
}
}
\description{
Returns information about the available schemes.
}
\examples{
## Get a table of available palettes
info()
}
\seealso{
Other color schemes: 
\code{\link{colour}()},
\code{\link{ramp}()}
}
\author{
N. Frerebeau
}
\concept{color schemes}
