% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/khroma-deprecated.R
\name{khroma-deprecated}
\alias{khroma-deprecated}
\alias{convert}
\alias{ramp}
\title{Deprecated Functions in dimensio}
\usage{
convert(x, mode)

ramp(palette, ...)
}
\description{
These functions still work but will be removed (defunct) in the next version.
}
\keyword{internal}
