% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/khroma-deprecated.R
\name{khroma-deprecated}
\alias{khroma-deprecated}
\alias{palette_size_range}
\title{Deprecated Functions in dimensio}
\usage{
palette_size_range(range = c(1, 6), ...)
}
\description{
These functions still work but will be removed (defunct) in the next version.
}
\keyword{internal}
