% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/khroma.R
\docType{package}
\name{khroma-package}
\alias{khroma-package}
\alias{khroma}
\title{Colour Schemes for Archaeological Data Visualization}
\description{
This package provides Paul Tol's colour schemes and several thematic palettes
(geologic timescale, FAO soils, etc.) with scales for 'ggplot2'.
}
\details{
\tabular{ll}{
 \strong{Package:} \tab khroma \cr
 \strong{Type:} \tab Package \cr
 \strong{Version:} \tab 1.1.0 \cr
 \strong{Date:} \tab 2019-01-09 \cr
 \strong{License:} \tab GPL-3 \cr
 \strong{DOI:} \tab \href{https://doi.org/10.5281/zenodo.1472077}{10.5281/zenodo.1472077}
}
}
\author{
\strong{Full list of authors and contributors} (alphabetic order)
 \tabular{ll}{
  Nicolas Frerebeau \tab \emph{IRAMAT-CRP2A, Université Bordeaux Montaigne, France} \cr
  Brice Lebrun \tab \emph{CEREGE, Aix-Marseille Université, France}
 }

\strong{Bug reporting}

- \url{https://github.com/nfrerebeau/khroma/issues}

\strong{Package maintainer}

Nicolas Frerebeau\cr
\email{nicolas.frerebeau@u-bordeaux-montaigne.fr}

IRAMAT-CRP2A (UMR 5060)\cr
Maison de l'Archéologie\cr
Universite Bordeaux Montaigne\cr
F-33607 Pessac cedex\cr
France
}
