% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/khroma-package.R
\docType{package}
\name{khroma-package}
\alias{khroma-package}
\alias{_PACKAGE}
\alias{khroma}
\title{khroma: Colour Schemes for Scientific Data Visualization}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

An implementation of Paul Tol's colour schemes for
    use with 'graphics' or 'ggplot2'. These schemes are ready for each
    type of data (qualitative, diverging or sequential), with colours that
    are distinct for all people, including colour-blind readers. This
    package provides tools to simulate colour-blindness and to test how
    well the colours of any palette are identifiable. Several scientific
    thematic schemes (geologic timescale, land cover, FAO soils, etc.) are
    also implemented.
}
\details{
\tabular{ll}{
\strong{Package:} \tab khroma \cr
\strong{Type:} \tab Package \cr
\strong{Version:} \tab 1.4.1.9000 \cr
\strong{License:} \tab GPL-3 \cr
}

\doi{10.5281/zenodo.1472077}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://khroma.tesselle.org}
  \item \url{https://github.com/tesselle/khroma}
  \item Report bugs at \url{https://github.com/tesselle/khroma/issues}
}

}
\author{
\strong{Full list of authors and contributors} (alphabetic order)
\tabular{ll}{
Vincent Arel-Bundock \tab \emph{Université de Montréal, Canada} \cr
Nicolas Frerebeau \tab \emph{Université Bordeaux Montaigne, France} \cr
Brice Lebrun \tab \emph{Université Bordeaux Montaigne, France}
}

\strong{Package maintainer}

Nicolas Frerebeau\cr
\email{nicolas.frerebeau@u-bordeaux-montaigne.fr}

IRAMAT-CRP2A (UMR 5060)\cr
Maison de l'Archéologie\cr
Université Bordeaux Montaigne\cr
F-33607 Pessac cedex\cr
France
}
\keyword{internal}
