% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/khroma-package.R
\docType{package}
\name{khroma-package}
\alias{khroma-package}
\alias{khroma}
\title{khroma: Colour Schemes for Scientific Data Visualization}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Color schemes ready for each type of data (qualitative, diverging or sequential), with colors that are distinct for all people, including color-blind readers. This package provides an implementation of Paul Tol (2018) and Fabio Crameri (2018) \doi{10.5194/gmd-11-2541-2018} color schemes for use with 'graphics' or 'ggplot2'. It provides tools to simulate color-blindness and to test how well the colors of any palette are identifiable. Several scientific thematic schemes (geologic timescale, land cover, FAO soils, etc.) are also implemented.
}
\details{
\tabular{ll}{
\strong{Package:} \tab khroma \cr
\strong{Type:} \tab Package \cr
\strong{Version:} \tab 1.15.0 \cr
\strong{License:} \tab GPL-3 \cr
\strong{Zenodo:} \tab \doi{10.5281/zenodo.1472077} \cr
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://codeberg.org/tesselle/khroma}
  \item \url{https://packages.tesselle.org/khroma/}
  \item Report bugs at \url{https://codeberg.org/tesselle/khroma/issues}
}

}
\author{
\strong{Full list of authors and contributors} (alphabetic order)
\tabular{ll}{
Vincent Arel-Bundock \tab \emph{Université de Montréal, Canada} \cr
Nicolas Frerebeau \tab \emph{Université Bordeaux Montaigne, France} \cr
Brice Lebrun \tab \emph{Université Bordeaux Montaigne, France} \cr
Ulrik Stervbo \tab \emph{Ruhr-Universität Bochum}
}

\strong{Package maintainer}

Nicolas Frerebeau\cr
\email{nicolas.frerebeau@u-bordeaux-montaigne.fr}

Archéosciences Bordeaux (UMR 6034)\cr
Maison de l'Archéologie\cr
Université Bordeaux Montaigne\cr
F-33607 Pessac cedex\cr
France
}
\keyword{internal}
