% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{palette_color_discrete}
\alias{palette_color_discrete}
\alias{palette_colour_discrete}
\title{Color Mapping (discrete)}
\usage{
palette_color_discrete(
  colors = NULL,
  domain = NULL,
  ordered = FALSE,
  missing = "#DDDDDD"
)

palette_colour_discrete(
  colors = NULL,
  domain = NULL,
  ordered = FALSE,
  missing = "#DDDDDD"
)
}
\arguments{
\item{colors}{A vector of colors that values will be mapped to. If \code{NULL}
(the default), uses \emph{viridis} (see \code{\link[grDevices:palettes]{grDevices::hcl.colors()}}).}

\item{domain}{A vector of categorical data specifying the possible values
that can be mapped.}

\item{ordered}{A \code{\link{logical}} scalar: should the levels be treated as already
in the correct order?}

\item{missing}{The color to return for \code{NA} values.}
}
\value{
A palette \code{\link{function}} that when called with a single argument
(a vector of categorical values) returns a \code{\link{character}} vector of colors.
}
\description{
Maps categorical values to colors.
}
\examples{
## Build palette functions
bright <- c(versicolor = "#4477AA", virginica = "#EE6677", setosa = "#228833")
ramp_bright <- palette_color_discrete(colors = bright)

## Plot
plot(
  x = iris$Petal.Length,
  y = iris$Sepal.Length,
  pch = 16,
  col = ramp_bright(iris$Species),
  xlab = "Petal length",
  ylab = "Sepal length",
  panel.first = grid(),
  las = 1
)
legend("topleft", legend = names(bright), col = bright, pch = 16)
}
\seealso{
Other palettes: 
\code{\link{palette_color_continuous}()},
\code{\link{palette_color_picker}()},
\code{\link{palette_shape}()},
\code{\link{palette_size_range}()}
}
\concept{palettes}
