% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{palette_size_range}
\alias{palette_size_range}
\title{Symbol Size Mapping}
\usage{
palette_size_range(range = c(1, 6), ...)
}
\arguments{
\item{range}{A length-two \code{\link{numeric}} vector giving range of possible sizes
(greater than 0).}

\item{...}{Currently not used.}
}
\value{
A palette \code{\link{function}} that when called with a single argument
(a \code{\link{numeric}} vector of continuous values) returns a \code{\link{numeric}} vector
giving the amount by which plotting text and symbols should be magnified
relative to the default.
}
\description{
Symbol Size Mapping
}
\seealso{
Other palettes: 
\code{\link{palette_color_continuous}()},
\code{\link{palette_color_discrete}()},
\code{\link{palette_color_picker}()},
\code{\link{palette_shape}()}
}
\concept{palettes}
