% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.R
\name{scale}
\alias{scale}
\title{Colour scale constructor}
\usage{
scale(aesthetics, scale_name, reverse = FALSE, midpoint = 0, ...)
}
\arguments{
\item{aesthetics}{The names of the aesthetics that this scale works with.}

\item{scale_name}{A \code{\link{character}} string giving the name of
the palette to be used (see \code{\link{colour}}).}

\item{reverse}{A \code{\link{logical}} scalar specifying if the resulting
vector of colours should be reversed.}

\item{midpoint}{A \code{\link{numeric}} value specifying the midpoint (in
data value) of the diverging scale (defaults to \eqn{0}).}

\item{...}{Further arguments passed to \code{\link[ggplot2]{discrete_scale}}
or \code{\link[ggplot2]{continuous_scale}}, used respectively for
qualitative data and diverging/sequential data.}
}
\description{
Builds discrete or continuous scale for 'ggplot2' according to the colour
 scheme used.
}
\author{
N. Frerebeau
}
\keyword{internal}
